Swapped floats for doubles

+++ src/main/kotlin/hr/fer/zemris/demo/4. Perspective.kt
@@ -21,35 +21,35 @@ fun main() {
     val canvas = display.canvas
 
     val viewPort = ScreenSpaceTransform(display.width, display.height)
-    val fovPerspectiveProjection = FovPerspectiveProjection(Math.toRadians(12.0).toFloat(), display.width.toFloat() / display.height.toFloat(), 0.1f, 500f)
+    val fovPerspectiveProjection = FovPerspectiveProjection(Math.toRadians(12.0), display.width.toDouble() / display.height.toDouble(), 0.1, 500.0)
 
     val cubeMesh = cubeVertices
     val pyramidMesh = pyramidVertices
     val diamondMesh = diamondVertices
 
-    var deltaRotCube1 = 0.02f
-    var deltaRotCube2 = 1f
-    var deltaRotCube3 = 0f
-    var deltaRotPyramid1 = 0.2f
-    var deltaRotPyramid2 = 0f
-    var deltaRotDiamond = 0f
+    var deltaRotCube1 = 0.02
+    var deltaRotCube2 = 1.0
+    var deltaRotCube3 = 0.0
+    var deltaRotPyramid1 = 0.2
+    var deltaRotPyramid2 = 0.0
+    var deltaRotDiamond = 0.0
     while (true) {
         canvas.clear(Color.BLACK)
 
-        val cubeModelMatrix1 = scaleMatrix(9f) * rotateXMatrix(deltaRotCube1) * rotateYMatrix(deltaRotCube1) * rotateZMatrix(deltaRotCube1) * translateMatrix(15f, 0f, -150f)
-        val cubeModelMatrix2 = scaleMatrix(5f) * rotateYMatrix(deltaRotCube2) * rotateZMatrix(deltaRotCube2) * translateMatrix(5f, 1f, -150f) * rotateZMatrix(deltaRotCube2/2)
-        val cubeModelMatrix3 = scaleXMatrix(5f) * scaleYMatrix(2f) * scaleMatrix(5f) * rotateYMatrix(deltaRotCube3) * translateMatrix(-20f, 5f, -250f)
-        val pyramidModelMatrix1 = scaleMatrix(5f) * rotateXMatrix(deltaRotPyramid1 + 0.5f) * rotateYMatrix(deltaRotPyramid1) * translateMatrix(0f, 0f, -150f)
-        val pyramidModelMatrix2 = scaleMatrix(5f) * rotateYMatrix(deltaRotPyramid2) * translateMatrix(-10f, -10f, -150f) * rotateZMatrix(deltaRotPyramid2)
-        val diamondModelMatrix = scaleYMatrix(2f) * scaleMatrix(4f) * rotateYMatrix(deltaRotDiamond) * rotateZMatrix(deltaRotDiamond) * translateMatrix(20f, -12f, -200f)
+        val cubeModelMatrix1 = scaleMatrix(9.0) * rotateXMatrix(deltaRotCube1) * rotateYMatrix(deltaRotCube1) * rotateZMatrix(deltaRotCube1) * translateMatrix(15.0, 0.0, -150.0)
+        val cubeModelMatrix2 = scaleMatrix(5.0) * rotateYMatrix(deltaRotCube2) * rotateZMatrix(deltaRotCube2) * translateMatrix(5.0, 1.0, -150.0) * rotateZMatrix(deltaRotCube2/2)
+        val cubeModelMatrix3 = scaleXMatrix(5.0) * scaleYMatrix(2.0) * scaleMatrix(5.0) * rotateYMatrix(deltaRotCube3) * translateMatrix(-20.0, 5.0, -250.0)
+        val pyramidModelMatrix1 = scaleMatrix(5.0) * rotateXMatrix(deltaRotPyramid1 + 0.5) * rotateYMatrix(deltaRotPyramid1) * translateMatrix(0.0, 0.0, -150.0)
+        val pyramidModelMatrix2 = scaleMatrix(5.0) * rotateYMatrix(deltaRotPyramid2) * translateMatrix(-10.0, -10.0, -150.0) * rotateZMatrix(deltaRotPyramid2)
+        val diamondModelMatrix = scaleYMatrix(2.0) * scaleMatrix(4.0) * rotateYMatrix(deltaRotDiamond) * rotateZMatrix(deltaRotDiamond) * translateMatrix(20.0, -12.0, -200.0)
 
 
-        deltaRotCube1 += 0.01f
-        deltaRotCube2 += 0.05f
-        deltaRotCube3 += 0.005f
-        deltaRotPyramid1 += 0.02f
-        deltaRotPyramid2 += 0.02f
-        deltaRotDiamond += 0.05f
+        deltaRotCube1 += 0.01
+        deltaRotCube2 += 0.05
+        deltaRotCube3 += 0.005
+        deltaRotPyramid1 += 0.02
+        deltaRotPyramid2 += 0.02
+        deltaRotDiamond += 0.05
 
         renderMesh(cubeMesh, canvas, Color.MAGENTA, cubeModelMatrix1, fovPerspectiveProjection.projectionMatrix, viewPort.viewPortMatrix)
         renderMesh(cubeMesh, canvas, Color.WHITE, cubeModelMatrix2, fovPerspectiveProjection.projectionMatrix, viewPort.viewPortMatrix)
@@ -75,4 +75,4 @@ private fun renderMesh(vertices: List<Vector>, canvas : Canvas, color: Color, mo
 }
 
 private fun vectorToPoint(v1: Vector, modelMatrix: Matrix, perspective: Matrix, viewPort: Matrix): Point =
-    ((v1.toMatrix(Vector.ToMatrix.ROW) * modelMatrix * perspective).let { it * (1f / it[0, 3]) } * viewPort).toVector().let { Point(ceil(it[0]).toInt(), ceil(it[1]).toInt()) }
+    ((v1.toMatrix(Vector.ToMatrix.ROW) * modelMatrix * perspective).let { it * (1.0 / it[0, 3]) } * viewPort).toVector().let { Point(ceil(it[0]).toInt(), ceil(it[1]).toInt()) }
