+++ src/main/kotlin/hr/fer/zemris/demo/7. Obj loading - temple.kt
@@ -20,7 +20,7 @@ fun main() {
     val renderer = FeRenderer(
         Display(1600, 900, "Object Loading Temple"),
         CameraImpl(position = vector(0, 0, 20), target = vector(0, 0, -20f)),
-        FovPerspectiveProjection(Math.toRadians(45.0).toFloat(), 16f/9f, 0.1f, 500f),
+        FovPerspectiveProjection(Math.toRadians(45.0), 16.0/9.0, 0.1, 500.0),
         ScreenSpaceTransform(1600, 900)
     ).apply {
         keyEventProcessor = { keyEvent ->
@@ -55,7 +55,7 @@ fun main() {
         RenderType.DRAW
     )
 
-    var deltaRot = 0f
+    var deltaRot = 0.0
 
     var fpsCounter = 0
     var time = System.currentTimeMillis()
@@ -63,7 +63,7 @@ fun main() {
         renderer.clearDisplay()
         renderer.processKeyEvents()
 
-        temple.modelViewTransform = scaleMatrix(20f) * rotateYMatrix(deltaRot) * translateMatrix(0f, 0f, -20f)
+        temple.modelViewTransform = scaleMatrix(20.0) * rotateYMatrix(deltaRot) * translateMatrix(0.0, 0.0, -20.0)
         deltaRot += 0.05f
 
         renderer.render(temple)
